refactor: extract CLI code into focused modules for better maintainability
- Extract 1796-line monolithic cli.py into 4 focused modules:
  - cli_visualization.py: dependency graphs and execution display
  - cli_config.py: configuration and provider management
  - cli_auth.py: authentication command implementations
  - cli_commands.py: main CLI command implementations
- Update cli.py to serve as clean command interface with imports
- Fix all test import paths to match new module structure
- Maintain 100% functionality with all 226 tests passing
- Remove unused code and resolve linting issues
- Improve code organization with clear separation of concerns
